+++ app/Http/Controllers/NoticeController.php
@@ -3,6 +3,7 @@
 namespace App\Http\Controllers;
 
 use App\Models\Notice;
+use App\Models\Teacher;
 use Illuminate\Http\Request;
 
 class NoticeController extends Controller
@@ -28,13 +29,20 @@ class NoticeController extends Controller
             't_id'=>'required',
         ]);
 
+        $id = $request->session()->get('user_id');
+        $teacher = Teacher::where('t_id', $id)->first();
+        if(!$teacher->status){
+            return back();
+        }
+
         Notice::create($request->all());
         return redirect('/teacher/notice');
     }
 
-    public function show(Notice $notice)
+    public function show(Notice $notice, Request $request)
     {
-        $notices= Notice::paginate(5);
+        $id = $request->session()->get('user_id');
+        $notices= Notice::where('t_id',$id)->orderBy('updated_at', 'desc')->paginate(5);
         return view('teacher.noticeEdit',['notices'=>$notices, 'notice'=>$notice]);
     }
     public function edit(Notice $notice)
@@ -45,19 +53,28 @@ class NoticeController extends Controller
     {
         $request['t_id']=$request->session()->get('user_id');
 
-        // dd($notice);
-
         $request->validate([
             'notice_description'=>'required',
             't_id'=>'required'
         ]);
 
+        $id = $request->session()->get('user_id');
+        $teacher = Teacher::where('t_id', $id)->first();
+        if(!$teacher->status){
+            return back();
+        }
+
         $notice->update($request->all());
         return redirect('/teacher/notice');
     }
 
-    public function destroy(Notice $notice)
+    public function destroy(Notice $notice, Request $request)
     {
+        $id = $request->session()->get('user_id');
+        $teacher = Teacher::where('t_id', $id)->first();
+        if(!$teacher->status){
+            return back();
+        }
         $notice->delete();
         return redirect('/teacher/notice');
     }
